rename directories to lowercase
