+++ apps/src/tests/TestSafeAreaViewIOS/bottom-tabs/BottomTabsComponent.tsx
