+++ apps/src/tests/TestSafeAreaViewIOS/bottom-tabs/index.tsx
