+++ apps/src/tests/TestSafeAreaViewIOS/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import StackV4SAVExample from './StackV4';
+import StackV4SAVExample from './stack-v4';
 // import BottomTabsSAVExample from './BottomTabs';
 // import SplitViewSAVExample from './SplitView';
